drivers: modem: hl78xx: split auxiliary GNSS parsers
Move auxiliary GNSS NMEA parsing helpers into a dedicated parser module.

Add hl78xx_gnss_aux_parse.c and hl78xx_gnss_aux_parse.h for parsing GSA,
GST, and PSEPU argument lists into HL78XX auxiliary GNSS data. Keep UTC
and optional milli-unit parsing private to that module.

Leave modem chat callback handling in hl78xx_gnss_parsers.c, where the
callbacks can update runtime GNSS state and publish auxiliary data. This
keeps the parsing logic testable without coupling it to modem chat
callback flow.

Build the new auxiliary parser module when HL78XX GNSS support is
enabled.
